URL and MySQL injection protection

+++ admin/includes/view_all_users.php
@@ -80,9 +80,13 @@ while ($row = mysqli_fetch_assoc($select_users)) {
   }
 
   if(isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_user_query = mysqli_query($connection, $query); 
-    header("Location: users.php");
+    if(isset($_SESSION['user_role'])) {
+      if($_SESSION['user_role'] == 'admin') {
+        $the_user_id = mysqli_real_escape_string($connection,$_GET['delete']);
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        $delete_user_query = mysqli_query($connection, $query); 
+        header("Location: users.php");
+      }
+    }
   }
 ?>
